% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vulnpryr.R
\name{load_vulndb}
\alias{load_vulndb}
\title{Load VulnDB extract}
\usage{
load_vulndb(vulndb_file)
}
\arguments{
\item{vulndb_file}{Path to VulnDB extract}
}
\description{
Load VulnDB extract
}

