% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viqtest.R
\name{set_vulndb}
\alias{set_vulndb}
\title{Set the vulnerability attributes database for use by vulnpryr}
\usage{
set_vulndb(vulndb)
}
\arguments{
\item{vulndb}{Dataframe of vulnerability attributes}
}
\description{
Set the vulnerability attributes database for use by vulnpryr
}

