% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viqtest.R
\docType{data}
\name{vulndb}
\alias{vulndb}
\title{Sample vulnerability attributes data}
\format{A data frame with 10 rows and 9 variables}
\usage{
data(vulndb)
}
\description{
A dataset of _dummy_ vulnerabilities attributes to demonstrate the
functionality and use of vulnpryr. The variables are as follows:
}
\details{
\itemize{
\item cve_id. CVE ID of the vulnerability
\item cvss_score. base CVSS  score (0 - 10)
\item msp. number of entries for the vulnerability in Metasploit (0+)
\item edb. number of entries for the vulnerability in Exploit DB (0+)
\item public_exploit. number of times the vulnerability has a known public vulnerability
\item private_exploit. number of times the vulnerability has a known private vulnerability
\item network_vector. number of times the vulnerability has a known network attack vector
\item impact_integrity. number of times the vulnerability has a known impact on integrity
\item impact_confidentiality. number of times the vulnerability has a known impact on confidentiality
}
}
\keyword{vulndb}

