% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viqtest.R
\name{vulnpryr}
\alias{vulnpryr}
\title{Rescale vulnerabilities.}
\usage{
vulnpryr(cve_id, cvss_base, avg_cvss_score = 6.2, msp_factor = 2.5,
  edb_factor = 1.5, private_exploit_factor = 0.5,
  network_vector_factor = 2, impact_factor = 3)
}
\arguments{
\item{cve_id}{ID of the CVE in question}

\item{cvss_base}{The current CVSS rating of the vuln in question}

\item{avg_cvss_score}{Mean CVSS score of the population}

\item{msp_factor}{Amount to adjust CVSS if Metasploit module is present}

\item{edb_factor}{Amount to adjust CVSS if ExploitDB is present}

\item{private_exploit_factor}{Factor if private exploit exists}

\item{network_vector_factor}{Amount to adjust if not a network vuln}

\item{impact_factor}{Amount to adjust if impact is not confidentiality}
}
\value{
Dataframe with the adjusted vuln
}
\description{
Rescale vulnerabilities.
}
\examples{
data(vulndb)
set_vulndb(vulndb)
vulnpryr("CVE-2013-2899", 5)
}

