% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viqtest.R
\name{load_vulndb}
\alias{load_vulndb}
\title{Load the vulnerability attributes extract}
\usage{
load_vulndb(vulndb_file)
}
\arguments{
\item{vulndb_file}{Path to vulnerability attributes extract}
}
\description{
Load the vulnerability attributes extract
}

