% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vulnpryr.R
\docType{package}
\name{vulnpryr}
\alias{vulnpryr}
\alias{vulnpryr-package}
\title{vulnpryr: A package for prying additional utility into your CVSS scores.}
\usage{
vulnpryr(cve_id, cvss_base, vulndb, avg_cvss_score = 6.2, msp_factor = 2.5,
  edb_factor = 1.5, private_exploit_factor = 0.5,
  network_vector_factor = 2, impact_factor = 3)
}
\arguments{
\item{cve_id}{ID of the CVE in question}

\item{cvss_base}{The current CVSS rating of the vuln in question}

\item{vulndb}{Dataframe of the vulndb extract}

\item{avg_cvss_score}{Mean CVSS score of the population}

\item{msp_factor}{Amount to adjust CVSS if Metasploit module is present}

\item{edb_factor}{Amount to adjust CVSS if ExploitDB is present}

\item{private_exploit_factor}{Factor if private exploit exists}

\item{network_vector_factor}{Amount to adjust if not a network vuln}

\item{impact_factor}{Amount to adjust if impact is not confidentiality}
}
\value{
Dataframe with the adjusted vuln
}
\description{
vulnpryr: A package for prying additional utility into your CVSS scores.

Add together two numbers.
}
\section{vulnpryr functions}{

The vulnpryer functions ...
}

